feat: Update Google Charts initialization, summernote initialization, logout functionality, navigation links, and blog post links

+++ admin/functions.php
@@ -81,7 +81,7 @@ function displayAllPosts()
     echo "<td>{$post_date}</td>";
     echo "<td><img width='100' src='../images/{$post_image}' alt='image'></td>";
     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}' >Edit</a></td>";
-    echo "<td><a href='posts.php?delete={$post_id}' >Delete</a></td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?')\" href='posts.php?delete={$post_id}' >Delete</a></td>";
 
 
   }
@@ -115,7 +115,7 @@ function displayAllUsers()
     echo "<td>{$user_role}</td>";
     echo "<td>{$randSalt}</td>";
     echo "<td><a href='users.php?source=edit_user&u_id={$user_id}' >Edit</a></td>";
-    echo "<td><a href='users.php?delete={$user_id}' >Delete</a></td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete?')\" href='users.php?delete={$user_id}' >Delete</a></td>";
 
 
   }
